test_requirements.txt
black
bpython
codecov
ddt
debugpy
django-debug-toolbar
factory-boy==3.2.0
faker
flaky
freezegun==0.3.14
hypothesis==4.23.9
ipdb
isort
nplusone
pdbpp
pip-tools
pylint==2.6.0
pylint-django==2.4.4
pytest
pytest-cov
pytest-env==0.6.2
pytest-django
pytest-mock
pytest-lazy-fixture==0.6.3
responses
semantic-version
wagtail-factories
mitol-django-openedx==1.0.0
